Removed DailyStats table. Fixed missing key prop error with run table rows. Better component structure.
