+++ src/components/Runs/AllRuns.jsx
@@ -4,25 +4,21 @@ import { StateContext } from '../../context/StateContext'
 import actions from '../../reducers/actions'
 import { APIv1 } from '../../api'
 
-// Third Party UI components
-import Switch from 'react-switch'
-
-// Colors - should probably pull this from Tailwind
-const eggplant700 = '#7A2A46'
-
 // Components
 import StravaImport from './StravaImport'
 import RunTableHeaders from './RunTableHeaders'
-import RunTableRows from './RunTableRows'
-import DailyStatsTableHeaders from './DailyStatsTableHeaders'
-import DailyStatsTableRows from './DailyStatsTableRows'
+import RunTableBody from './RunTableBody'
 
-// Should the group by date toggle be on or off?
-const getInitialGroupByState = () => {
-  if (localStorage.getItem('groupByDate') == null) {
-    return true // Want runs grouped by date for new users. Might be a dumb setting but who care.
-  }
-  return JSON.parse(localStorage.getItem('groupByDate'))
+// Tell Runlog API to talk to Strava and get the most recent run activities
+const getRecentStravaRuns = () => {
+  // Go get the full user information
+  APIv1.get('/strava/runs')
+    .then((response) => {
+      return response.data
+    })
+    .catch((error) => {
+      console.error(error)
+    })
 }
 
 const AllRuns = () => {
@@ -30,25 +26,11 @@ const AllRuns = () => {
   const auth = useContext(AuthContext)[0]
   const [state, dispatch] = useContext(StateContext)
   const hasStravaAccount = auth.user && auth.user.hasStravaAuth
-  const [groupByDate, setGroupByDate] = useState(getInitialGroupByState())
-
-  if (
-    state == null ||
-    state.runs === null ||
-    state.runs.byId === null ||
-    state.dailyStats.byId === null
-  ) {
-    return <div className='AllRuns w-full px-4 pb-4'></div>
-  }
 
   // When this component is loaded, go get the user's Runlog runs (all of them at once)
   useEffect(() => {
     dispatch({
       type: actions.GET_ALL_RUNS__START,
-    })
-
-    dispatch({
-      type: actions.GET_ALL_DAILY_STATS__START,
     })
 
     APIv1.get('/runs')
@@ -64,76 +46,20 @@ const AllRuns = () => {
           error: error,
         })
       })
-
-    APIv1.get('/dailyStats')
-      .then((response) => {
-        dispatch({
-          type: actions.GET_ALL_DAILY_STATS__SUCCESS,
-          dailyStats: response.data,
-        })
-      })
-      .catch((error) => {
-        dispatch({
-          type: actions.GET_ALL_DAILY_STATS__ERROR,
-          error: error,
-        })
-      })
   }, [])
 
-  useState()
-
-  // Tell Runlog API to talk to Strava and get the most recent run activities
-  const getRecentStravaRuns = () => {
-    // Go get the full user information
-    APIv1.get('/strava/runs')
-      .then((response) => {
-        setStravaRuns(response.data)
-      })
-      .catch((error) => {
-        console.error(error)
-      })
+  if (state == null || state.runs === null || state.runs.byId === null) {
+    return null
   }
 
   return (
     <div className='AllRuns w-full pb-4 space-y-4 overflow-auto w-full'>
-      <div className='mx-4'>
-        <label className='inline-flex items-center space-x-2'>
-          <span>Group by date</span>
-          <Switch
-            onChange={() => {
-              localStorage.setItem('groupByDate', !groupByDate)
-              setGroupByDate(!groupByDate)
-            }}
-            checked={groupByDate}
-            uncheckedIcon={false}
-            checkedIcon={false}
-            activeBoxShadow={null}
-            onColor={eggplant700}
-            height={20}
-            width={40}
-          />
-        </label>
-      </div>
+      <h1 className='mx-4 text-2xl'>All Runs</h1>
 
-      {groupByDate && (
-        <section className='overflow-scroll grid grid-cols-daily-stats-page'>
-          <DailyStatsTableHeaders />
-          <DailyStatsTableRows
-            dailyStats={state.dailyStats.byId}
-            isLoading={state.dailyStats.isLoading}
-          />
-        </section>
-      )}
-
-      {!groupByDate && (
-        <section className='overflow-scroll grid grid-cols-runs-page'>
-          <RunTableHeaders />
-          <RunTableRows
-            runs={state.runs.byId}
-            isLoading={state.runs.isLoading}
-          />
-        </section>
-      )}
+      <section className='overflow-scroll grid grid-cols-runs-page'>
+        <RunTableHeaders />
+        <RunTableBody runs={state.runs.byId} isLoading={state.runs.isLoading} />
+      </section>
 
       {hasStravaAccount && (
         <>
@@ -150,7 +76,7 @@ const AllRuns = () => {
 
             <button
               className='px-4 py-2 border text-white border-gray-900 rounded bg-eggplant-700 hover:bg-eggplant-600 transition focus:outline-none'
-              onClick={getRecentStravaRuns}
+              onClick={setStravaRuns(getRecentStravaRuns)}
             >
               Get Strava Runs
             </button>
