+++ src/reducers/actions.js
@@ -22,11 +22,6 @@ export default {
   DELETE_RUN__SUCCESS: 'DELETE_RUN__SUCCESS',
   DELETE_RUN__ERROR: 'DELETE_RUN__ERROR',
 
-  // Daily Stats actions
-  GET_ALL_DAILY_STATS__START: 'GET_ALL_DAILY_STATS__START',
-  GET_ALL_DAILY_STATS__SUCCESS: 'GET_ALL_DAILY_STATS__SUCCESS',
-  GET_ALL_DAILY_STATS__ERROR: 'GET_ALL_DAILY_STATS__ERROR',
-
   // Training Plans actions
   GET_ALL_TRAINING__START: 'GET_ALL_TRAINING__START',
   GET_ALL_TRAINING__SUCCESS: 'GET_ALL_TRAINING__SUCCESS',
